Rename backend foundation PersistenceInfrastructure project to InfrastructureCore
