+++ account-management/Infrastructure/ApplicationDbContext.cs
@@ -2,7 +2,7 @@ using System.Diagnostics.CodeAnalysis;
 using Microsoft.EntityFrameworkCore;
 using PlatformPlatform.AccountManagement.Domain.Tenants;
 using PlatformPlatform.Foundation.DomainModeling.Entities;
-using PlatformPlatform.Foundation.PersistenceInfrastructure.EntityFramework;
+using PlatformPlatform.Foundation.InfrastructureCore.EntityFramework;
 
 namespace PlatformPlatform.AccountManagement.Infrastructure;
