+++ backend-foundation/AspNetCoreUtils/AspNetCoreUtils.csproj
@@ -19,7 +19,7 @@
 
     <ItemGroup>
         <ProjectReference Include="..\DomainModeling\DomainModeling.csproj"/>
-        <ProjectReference Include="..\PersistenceInfrastructure\PersistenceInfrastructure.csproj"/>
+        <ProjectReference Include="..\InfrastructureCore\InfrastructureCore.csproj"/>
     </ItemGroup>
 
 </Project>
