+++ backend-foundation/DomainModeling/Validation/ValidationUtils.cs
@@ -6,10 +6,8 @@ namespace PlatformPlatform.Foundation.DomainModeling.Validation;
 ///     The ValidationUtils class is a static utility class that provides methods for validating various types of input
 ///     data such as strings, phone numbers, and email addresses. These methods are used for validating the data of
 ///     domain entities and commands. The methods return a <see cref="ValidationResult" /> object, which is used to
-///     determine if
-///     the input is valid or not. If the input is not valid, the <see cref="ValidationResult" /> object contains an error
-///     message
-///     that can be used to inform the user of the error.
+///     determine if the input is valid or not. If the input is not valid, the <see cref="ValidationResult" /> object
+///     contains an error message that can be used to inform the user of the error.
 /// </summary>
 public static partial class ValidationUtils
 {
