+++ backend-foundation/InfrastructureCore/InfrastructureCore.csproj
@@ -4,8 +4,8 @@
         <TargetFramework>net7.0</TargetFramework>
         <ImplicitUsings>enable</ImplicitUsings>
         <Nullable>enable</Nullable>
-        <RootNamespace>PlatformPlatform.Foundation.PersistenceInfrastructure</RootNamespace>
-        <AssemblyName>PlatformPlatform.Foundation.PersistenceInfrastructure</AssemblyName>
+        <RootNamespace>PlatformPlatform.Foundation.InfrastructureCore</RootNamespace>
+        <AssemblyName>PlatformPlatform.Foundation.InfrastructureCore</AssemblyName>
     </PropertyGroup>
 
     <ItemGroup>
