+++ backend-foundation/Tests/InfrastructureCore/Persistence/UnitOfWorkTests.cs
@@ -1,11 +1,11 @@
 using FluentAssertions;
 using Microsoft.EntityFrameworkCore;
 using PlatformPlatform.Foundation.DomainModeling.Persistence;
-using PlatformPlatform.Foundation.PersistenceInfrastructure.Persistence;
+using PlatformPlatform.Foundation.InfrastructureCore.Persistence;
 using PlatformPlatform.Foundation.Tests.TestEntities;
 using Xunit;
 
-namespace PlatformPlatform.Foundation.Tests.PersistenceInfrastructure.Persistence;
+namespace PlatformPlatform.Foundation.Tests.InfrastructureCore.Persistence;
 
 public sealed class UnitOfWorkTests : IDisposable
 {
